# src/main/java/com/pes/doacao_ms/configs/swagger/doacao.yaml
...
openapi: 3.0.1
info:
  title: API de Doação
  description: API para os serviços de doação e doadores
  version: 1.0.0
servers:
- url: http://localhost:8083
  description: Generated server url
paths:
  "/v1/doador":
    get:
      tags:
      - doador-controller
      summary: Get a Doador.
      description: Return a specific Doador by Email.
      operationId: get
      parameters:
      - name: emailRequest
        in: query
        required: true
        schema:
          "$ref": "#/components/schemas/EmailRequest"
      responses:
        '200':
          description: Doador found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DoadorResponse"
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                type: string
                example:
                  error: Invalid request body
        '404':
          description: Doador not found
          content:
            application/json:
              schema:
                type: string
                example:
                  error: Doador not found
    post:
      tags:
      - doador-controller
      summary: Create a new Doador.
      description: Create a new Doador using the given request.
      operationId: include
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/IncludeDoadorRequest"
        required: true
      responses:
        '201':
          description: Doador created
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DoadorIdResponse"
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                type: string
                example:
                  error:
                  - Name is required
                  - Document is required
...
components:
  schemas:
    IncludeDoadorRequest:
      required:
      - document
      - email
      - name
      - phone
      type: object
      properties:
        name:
          type: string
        document:
          type: string
        email:
          type: string
        phone:
          type: string
    DoadorIdResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
...
    EmailRequest:
      required:
      - email
      type: object
      properties:
        email:
          type: string
    DoadorResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        document:
          type: string
        email:
          type: string
        phone:
          type: string
        donationsNumber:
          type: integer
          format: int64